code changes

+++ src/main/java/com/intellifleet/dto/InstrumentPacketForRedisDTO.java
@@ -2,14 +2,16 @@ package com.intellifleet.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record InstrumentPacketForRedisDTO(
         String instrumentId,
         String resources,
         String tmsInstrument,
 
-        Double latitude,
-        Double longitude
+        BigDecimal latitude,
+        BigDecimal longitude
 
 ) {
 }
